ml-image-detector/requirements.txt
flask
pillow
numpy